# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API E-Commerce UAS",
        "description": "Dokumentasi lengkap API untuk integrasi Frontend Mobile (Flutter).",
        "contact": {
            "email": "hifadlibilal@sistekin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products (Public)"
                ],
                "summary": "Lihat Semua Produk (Untuk Home Buyer)",
                "description": "Menampilkan semua produk terbaru beserta data Sellernya.",
                "operationId": "c1ff6f862214e1896c59cfe0491ce0e8",
                "responses": {
                    "200": {
                        "description": "List Semua Produk",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "price": {
                                                        "type": "integer"
                                                    },
                                                    "image": {
                                                        "description": "Path gambar (tambahkan base_url/storage/ di frontend)",
                                                        "type": "string"
                                                    },
                                                    "seller": {
                                                        "properties": {
                                                            "name": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products (Seller)"
                ],
                "summary": "Tambah Produk Baru",
                "operationId": "399770d33fa4672f1216707d0e236763",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Laptop Gaming"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 15000000
                                    },
                                    "weight": {
                                        "description": "Berat dalam Gram (Penting buat Ongkir)",
                                        "type": "integer",
                                        "example": 2000
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Laptop spek dewa"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produk Berhasil Dibuat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/seller/products": {
            "get": {
                "tags": [
                    "Products (Seller)"
                ],
                "summary": "Lihat Produk Saya (Manage Produk)",
                "operationId": "308af00e08471dcd0b5381d8400c1e8d",
                "responses": {
                    "200": {
                        "description": "List Produk milik Seller yang sedang login"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/{id}": {
            "post": {
                "tags": [
                    "Products (Seller)"
                ],
                "summary": "Update Produk (Termasuk Gambar)",
                "description": "Gunakan method POST dengan body _method=PUT agar bisa upload file di Laravel",
                "operationId": "1c52fc58e33d71677ba74cb346470913",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "description": "Wajib diisi PUT",
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "integer"
                                    },
                                    "weight": {
                                        "type": "integer"
                                    },
                                    "stock": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produk Berhasil Diupdate"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products (Seller)"
                ],
                "summary": "Hapus Produk",
                "operationId": "fe34b364a3f192305598b65d04d06b5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produk Berhasil Dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan Token Bearer di sini untuk mengakses endpoint yang dilindungi.",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}